minor: Build specific jjversion commit (#9)
In particular, this builds jjversion v0.3.0,
which notably fixes the jjversion module path,
adding /src at the end. That change was
necessary in order to get the specific version,
because there was an error building jjversion:

module declares its path as: github.com/jjliggett/jjversion
  but was required as: github.com/jjliggett/jjversion/src

Unfortunately, this change is a breaking change
for jjversion-action, so the previous release
will no longer work after the jjversion change
is merged.