patch: Update jjversion-gha-output to v0.3.11 (#41)
This primarily relates to the v0.3.11 release of jjversion-gha-output:
https://github.com/jjliggett/jjversion-gha-output/releases/tag/v0.3.11
as well as the most recent pull request for jjversion-gha-output:
https://github.com/jjliggett/jjversion-gha-output/pull/49

This commit follows 38a75c396614ca123a0757739ba4dbc858a4219e in
additional updates to dependencies of jjversion.

A full list of relevant changes can be seen on these pages:
- https://github.com/jjliggett/jjversion-gha-output/compare/v0.3.8...v0.3.11
- https://github.com/jjliggett/jjversion/compare/v0.5.20...v0.5.24

The most notable changes include the following 3 changes:
- https://github.com/jjliggett/jjversion/pull/181
- https://github.com/jjliggett/jjversion/pull/184
- https://github.com/jjliggett/jjversion/pull/187

Beyond updating dependencies, this commit also includes a
small change to the GitHub workflow pipeline. The version
of jjversion-gha-output now is stored within a workflow
step environment variable, which reduces the quantity of
changes needed when new versions of jjversion-gha-output
are available.

Also, note, the prior commit, 38a75c3, had the wrong links
within its commit message. The links were pasted in error,
and the actual relevant pull requests for that commit
include the following:
- https://github.com/jjliggett/jjversion-action/pull/38
- https://github.com/jjliggett/jjversion-gha-output/pull/44
- https://github.com/jjliggett/jjversion/pull/174
- https://github.com/jjliggett/jjversion/pull/180